+++ pkg/plumbing/watch/encoding.go
@@ -12,20 +12,21 @@ import (
 	"github.com/warptools/warpforge/wfapi"
 )
 
-var encodeOptions dagjson.EncodeOptions = dagjson.EncodeOptions{
+var defaultEncodeOptions dagjson.EncodeOptions = dagjson.EncodeOptions{
 	EncodeLinks: false,
 	EncodeBytes: false,
 	MapSortMode: codec.MapSortMode_None,
 }
 
-var decodeOptions dagjson.DecodeOptions = dagjson.DecodeOptions{
+var defaultDecodeOptions dagjson.DecodeOptions = dagjson.DecodeOptions{
 	ParseLinks:         false,
 	ParseBytes:         false,
 	DontParseBeyondEnd: true, // This is critical for streaming over a socket
 }
 
+// encode wraps dagjson.Marshal with default options.
 func encode(n datamodel.Node, w io.Writer, opt rfmtjson.EncodeOptions) error {
-	err := dagjson.Marshal(n, rfmtjson.NewEncoder(w, opt), encodeOptions)
+	err := dagjson.Marshal(n, rfmtjson.NewEncoder(w, opt), defaultEncodeOptions)
 	if err != nil {
 		return serum.Error(wfapi.ECodeSerialization, serum.WithCause(err),
 			serum.WithMessageLiteral("watch encoder failed"),
@@ -34,6 +35,8 @@ func encode(n datamodel.Node, w io.Writer, opt rfmtjson.EncodeOptions) error {
 	return nil
 }
 
+// Encoder is a compact json encoder
+//
 // Errors:
 //
 //   - warpforge-error-serialization --
@@ -44,6 +47,8 @@ func Encoder(n datamodel.Node, w io.Writer) error {
 	})
 }
 
+// PrettyEncoder is a json encoder with line breaks and tab indentation.
+//
 // Errors:
 //
 //   - warpforge-error-serialization --
@@ -54,11 +59,13 @@ func PrettyEncoder(n datamodel.Node, w io.Writer) error {
 	})
 }
 
+// Decoder is a streaming JSON decoder which halts at the end of each object in the stream.
+//
 // Errors:
 //
 //   - warpforge-error-serialization --
 func Decoder(na datamodel.NodeAssembler, r io.Reader) error {
-	err := decodeOptions.Decode(na, r)
+	err := defaultDecodeOptions.Decode(na, r)
 	if err != nil {
 		return serum.Error(wfapi.ECodeSerialization, serum.WithCause(err),
 			serum.WithMessageLiteral("watch decoder failed"),
